+++ lib/authlogic/controller_adapters/abstract_adapter.rb
@@ -37,6 +37,10 @@ module Authlogic
         controller.params
       end
 
+      def headers
+        controller.headers
+      end
+
       def request
         controller.request
       end