+++ lib/authlogic/session/base.rb
@@ -236,6 +236,20 @@ module Authlogic
     #
     # You can modify all of this behavior with the Config sub module.
     #
+    # Headers
+    # =======
+    #
+    # This module is responsible for authenticating the user via headers, which requires
+    # setting HTTP header, for example, the following curl command:
+    #
+    #   curl -H "user_credentials: 4LiXF7FiGUppIPubBPey" https://www.domain.com
+    #
+    # Notice the token in the header parameter, this is a single access token. This header's
+    # method operates exactly as the params method.
+    #
+    # You can modify all of this behavior with the Config sub module.
+    #
+    # Disabled by default. To enable, set a non-nil headers_key.
     # Perishable Token
     # ================
     #
@@ -410,6 +424,7 @@ module Authlogic
 
       # `persist` callbacks, in order of priority
       persist :persist_by_params
+      persist :persist_by_headers
       persist :persist_by_cookie
       persist :persist_by_session
       persist :persist_by_http_auth, if: :persist_by_http_auth?
@@ -856,6 +871,24 @@ module Authlogic
           rw_config(:params_key, value, cookie_key)
         end
         alias params_key= params_key
+
+        # Works exactly like cookie_key, but for headers. So a user can login via
+        # headers just like a cookie or a session. Your URL would look like:
+        #
+        #   curl -H "user_credentials: 4LiXF7FiGUppIPubBPey" https://www.domain.com
+        #
+        # You can change the "user_credentials" key above with this
+        # configuration option. Keep in mind, just like cookie_key, if you
+        # supply an id the id will be appended to the front. Check out
+        # cookie_key for more details. Also checkout the "Single Access /
+        # Private Feeds Access" section in the README.
+        #
+        # * <tt>Default:</tt> nil
+        # * <tt>Accepts:</tt> String
+        def headers_key(value = nil)
+          rw_config(:headers_key, value)
+        end
+        alias headers_key= headers_key
 
         # Works exactly like login_field, but for the password instead. Returns
         # :password if a login_field exists.
@@ -1876,21 +1909,33 @@ module Authlogic
         controller.params[params_key]
       end
 
-      def params_enabled?
-        if !params_credentials || !klass.column_names.include?("single_access_token")
-          return false
-        end
-        if controller.responds_to_single_access_allowed?
-          return controller.single_access_allowed?
-        end
-        params_enabled_by_allowed_request_types?
+      def headers_credentials
+        # Setting headers_key to nil is the accepted way to disable
+        # single_access_token in headers.
+        return nil if headers_key.nil?
+        controller.headers[headers_key]
       end
 
-      def params_enabled_by_allowed_request_types?
-        case single_access_allowed_request_types
-        when Array
-          single_access_allowed_request_types.include?(controller.request_content_type) ||
-            single_access_allowed_request_types.include?(:all)
+      def params_enabled?
+        params_credentials && single_access_token_enabled?
+      end
+
+      def headers_enabled?
+        headers_credentials && single_access_token_enabled?
+      end
+
+      def single_access_token_enabled?
+        return false unless klass.column_names.include?("single_access_token")
+        return controller.single_access_allowed? if controller.responds_to_single_access_allowed?
+
+        single_access_token_allowed_by_request_type?
+      end
+
+      def single_access_token_allowed_by_request_type?
+        if single_access_allowed_request_types.is_a?(Array)
+          Set.new(single_access_allowed_request_types).intersect?(
+            Set[controller.request_content_type, :all]
+          )
         else
           %i[all any].include?(single_access_allowed_request_types)
         end
@@ -1898,6 +1943,10 @@ module Authlogic
 
       def params_key
         build_key(self.class.params_key)
+      end
+
+      def headers_key
+        build_key(self.class.headers_key)
       end
 
       def password_field
@@ -1919,10 +1968,17 @@ module Authlogic
       end
 
       def persist_by_params
-        return false unless params_enabled?
+        persist_by_single_access_token(params_credentials) if params_enabled?
+      end
+
+      def persist_by_headers
+        persist_by_single_access_token(headers_credentials) if headers_enabled?
+      end
+
+      def persist_by_single_access_token(credentials)
         self.unauthorized_record = search_for_record(
           "find_by_single_access_token",
-          params_credentials
+          credentials
         )
         self.single_access = valid?
       end