+++ lib/authlogic/test_case/mock_api_controller.rb
@@ -26,6 +26,10 @@ module Authlogic
         @params ||= {}
       end
 
+      def headers
+        @headers ||= {}
+      end
+
       def request
         @request ||= MockRequest.new(self)
       end